Added Support for Portable PDBs.   PerfView will now open either type and use that when looking up source code.
With this change you can do goto source on code that was built using portable PDBs and it will work.

Rename SymbolModule -> NativeSymbolModule  and
OpenSymbolFile -> OpenWindowsSymbolFile  which is what existing code should call.

Note that making gotoSource work with SourceLink attributed WINDOWS pdbs may not work yet.
That requires PerfView to read the new sourceLink information out of WINDOWS pdbs, which it does
not do today.  That is an independent feature.
