+++ src/PerfView/memory/PdbScopeMemoryGraph.cs
@@ -225,7 +225,7 @@ public class PdbScopeMemoryGraph : MemoryGraph
                     }
                 }
                 DebugWriteLine("Found pdb file " + pdbFilePath);
-                var module = symReader.OpenSymbolFile(pdbFilePath);
+                var module = symReader.OpenWindowsSymbolFile(pdbFilePath);
                 m_moduleMap = module.GetMergedAssembliesMap();
             }
         }
