+++ src/TraceEvent/TraceEvent.csproj
@@ -61,6 +61,7 @@
     <When Condition="'$(TargetFramework)' == 'net45'">
       <ItemGroup>
         <Reference Include="System.IO.Compression" />
+        <Reference Include="System.Net.Http" />
         <PackageReference Include="System.Runtime.InteropServices.RuntimeInformation" Version="4.3.0" />
       </ItemGroup>
     </When>
@@ -68,6 +69,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Diagnostics.Tracing.TraceEvent.SupportFiles" Version="$(MicrosoftDiagnosticsTracingTraceEventSupportFilesVersion)" />
+    <PackageReference Include="System.Reflection.Metadata" Version="1.5.0" />
   </ItemGroup>
 
   <ItemGroup>
